+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # RayPass Changelog
 
-## [Initial Version] - 2022-09-29
+## [Initial Version] - 2022-10-3
+
+- Launched RayPass 🎉